add xcmp call and move xcmp dmp to seperation config
